requirements.txt
requests>=2.25.0
feedparser>=6.0.0
PyGithub>=1.55
